requirements.txt
transformers
torch
google-genai
streamlit
Pillow